--- src/main/java/domain/piece/jumper/Jumper.java
@@ -1,10 +1,0 @@
-package domain.piece.jumper;
-
-import domain.piece.Camp;
-import domain.piece.Piece;
-
-public abstract class Jumper extends Piece {
-    protected Jumper(Camp camp) {
-        super(camp);
-    }
-}
