+++ src/test/java/domain/piece/pawn/PawnTest.java
@@ -12,7 +12,8 @@ import org.junit.jupiter.api.Test;
 import domain.board.Square;
 import domain.piece.Camp;
 import domain.piece.Piece;
-import domain.piece.empty.Empty;
+import domain.piece.nonslider.Empty;
+import domain.piece.nonslider.Pawn;
 
 class PawnTest {
